Make sure, that $$data will be available and will be an object, in places where we access it
Not checking `$$data` for type possibly leads to to accessing
properties of `null`, which is bad.

while `$$data` can only become `null` after `.destroy()` call, it is
still possible, that there might be attempts to access cache, most
notably - an async `onExpire` callback, which tries to store data in
cache after it was disabled.